+++ Lib/TSCompiler/tslib.js
@@ -116,6 +116,14 @@ var __spread = function () {
     return ar;
 };
 
+var __spreadArrays = function () {
+    for (var s = 0, i = 0, il = arguments.length; i < il; i++) s += arguments[i].length;
+    for (var r = Array(s), k = 0, i = 0; i < il; i++)
+        for (var a = arguments[i], j = 0, jl = a.length; j < jl; j++, k++)
+            r[k] = a[j];
+    return r;
+};
+
 var __await = function (v) {
     return this instanceof __await ? (this.v = v, this) : new __await(v);
 };